Android Studio kept crashing for me when attempting to load Compose previews. I had committed some work, but did not push it. Android Studio rolled back to before I commited the following: Added additional AboutNavigation function that gives users option to provide libraries themselves. Created directory for Library files. Created directory for PageIndicator files. Moved several functions/classes into their own files or existing files.
I unfortunately did not notice, so this commit is bigger than I wanted. The following is a new commit: Finished with UI of LibraryScreen, I ended up refactoring LibraryItem -> LibraryDetails and got it to work with both being a smaller Composable on AboutScreen and full screen on LibraryScreen. I still have to fix the styling for LibraryScreen.
