+++ androidutilities/src/main/java/com/heyzeusv/androidutilities/compose/ui/about/AboutNavigation.kt
@@ -11,13 +11,53 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import com.heyzeusv.androidutilities.compose.ui.library.LibraryPartyInfo
+import com.heyzeusv.androidutilities.compose.ui.library.LibraryScreen
+import com.heyzeusv.androidutilities.compose.ui.library.produceLibraryState
+import com.mikepenz.aboutlibraries.entity.Library
+
+@Composable
+fun AboutNavigation(
+    icon: @Composable () -> Unit = { },
+    title: String,
+    version: String,
+    info: List<String>,
+    separateByParty: Boolean = true,
+    aboutColors: AboutColors = AboutDefaults.aboutColors(),
+    aboutPadding: AboutPadding = AboutDefaults.aboutPadding(),
+    aboutDimensions: AboutDimensions = AboutDefaults.aboutDimensions(),
+    aboutTextStyles: AboutTextStyles = AboutDefaults.aboutTextStyles(),
+) {
+    val libraries by produceLibraryState(separateByParty = separateByParty)
+
+    AboutNavigation(
+        icon = icon,
+        title = title,
+        version = version,
+        info = info,
+        libraries = libraries,
+        aboutColors = aboutColors,
+        aboutPadding = aboutPadding,
+        aboutDimensions = aboutDimensions,
+        aboutTextStyles = aboutTextStyles,
+    )
+}
 
 @OptIn(ExperimentalSharedTransitionApi::class)
 @Composable
-fun AboutNavigation() {
+fun AboutNavigation(
+    icon: @Composable () -> Unit = { },
+    title: String,
+    version: String,
+    info: List<String>,
+    libraries: Map<LibraryPartyInfo, List<Library>>,
+    aboutColors: AboutColors = AboutDefaults.aboutColors(),
+    aboutPadding: AboutPadding = AboutDefaults.aboutPadding(),
+    aboutDimensions: AboutDimensions = AboutDefaults.aboutDimensions(),
+    aboutTextStyles: AboutTextStyles = AboutDefaults.aboutTextStyles(),
+) {
     SharedTransitionLayout {
         val navController = rememberNavController()
-        val libraries by produceLibraryState(true)
 
         // TODO: Try out both sharedElement and sharedBounds
         NavHost(
@@ -28,9 +68,18 @@ fun AboutNavigation() {
                 AboutScreen(
                     sharedTransitionScope = this@SharedTransitionLayout,
                     animatedContentScope = this,
+                    icon = icon,
+                    title = title,
+                    version = version,
+                    info = info,
+                    libraries = libraries,
                     libraryOnClick = { partyId, libraryId ->
                         navController.navigate("details/$partyId/$libraryId")
-                    }
+                    },
+                    colors = aboutColors,
+                    padding = aboutPadding,
+                    dimensions = aboutDimensions,
+                    textStyles = aboutTextStyles,
                 )
             }
             composable(
@@ -54,6 +103,10 @@ fun AboutNavigation() {
                             animatedVisibilityScope = this
                         ),
                     library = library,
+                    colors = aboutColors.libraryColors,
+                    padding = aboutPadding.libraryPadding,
+                    dimensions = aboutDimensions.libraryDimensions,
+                    textStyles = aboutTextStyles.libraryStyles,
                 )
             }
         }
