+++ app/src/main/java/com/heyzeusv/androidutilitieslibrary/util/FakeAboutData.kt
@@ -63,7 +63,7 @@ val fakeLongLibrary = Library(
     uniqueId = "com.google.fakeLibrary",
     artifactVersion = "1.0.0",
     name = "Fake Google Library Fake Google Library Fake Google Library",
-    description = "This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. This is a fake first party library used in previews. ",
+    description = lorem + lorem + lorem + lorem,
     website = "www.fake.com",
     developers = persistentListOf(fakeDeveloper, fakeDeveloper, fakeDeveloper, fakeDeveloper),
     organization = fakeOrganization,
